refactor: me ~> modes
Only a refactor, because leaving old path /me, aliased to /modes.

Removes silly FIXME note, while here.
